Simple hack to override mlr model from query string
This is a rather naive attempt to allow us to do a sanity
check on an MLR model before we roll it out. With the new
cirrusMLRModel query parameter we can upload a model to
elasticsearch and try a couple queries by specifying the
model name before we ship a config cange to enable the
model for everyone.

Change-Id: Id258c4ad295eab8cac543f14e9135a1bdb87533a