Use restify instead of lowlevel sockets
Probably overkill but dealing with request boudaries seemed a bit hard
to do when working directly on sockets.
Also I'm wondering if we could not use the filesystem as a locking
mechanism rather than creating a child process.
Not entirely sure but using restify I'm unable to keep the process
running when onComplete runs leading to errors preventing the unix
socket to be closed (added a hack to unlink it).

Change-Id: Ic0473887c4eb10d3dca3b7cfafabf3ab31654b1d