Port update_redirect_api.feature to nodejs
Mostly a straight copy from ruby. While most of the 'within ...' steps
could be removed unfortunately the one that waits for a redirect
converted to not a redirect couldn't be removed. This has to wait for
the previously redirected-to page to be updated. Rather than indirectly
waiting on search results added a new step that directly waits for the
redirect to be removed from the previously redirected to document.

Change-Id: I1febfdad9eac1f8e3577b545274378dd50ae5de0