Port update_weight_api.feature to nodejs
Replace 'within ...' calls with a new step that waits for incoming_links
to be appropriately updated. This makes our expectations explicit
instead of waiting on a secondary thing that hopefully updates based on
the first. Additionally to save some time (this test is pretty slow)
skip waiting on most of the edits and just do the final wait for
incoming_links.

I'm not really sure these tests even really need to be performing
searches, it looks like mostly they are checking that incoming links are
updated and counted appropriately.

While working up this patch i noticed multiple steps are all now
utilizing a 'wait for elasticsearch document to have some value' type
step so moved the implementation into stepHelpers and adjusted the uses
to all have similar wording and use the same implementation with
different check functions.

Change-Id: I20b13236e2139026d542de5e376392d6c5a67e47