Resolve redirect namespaces from source docs in fancy title results type
Rather than figuring out an appropriate namespace for redirects we
used the namespace of the document redirected to which is regularly
wrong, especially for 'shorthand' redirects on wikis such as
CAT:PROD on enwiki in NS_MAIN which redirects to Category:Proposed
Deletion (any many other similar shortcuts). Dig through the redirects
stored with the document and figure out what the likely namespace
of the document is.

This is a partial fix of T115756

Bug: T115756
Change-Id: I8de5c9d35ed709ee100cc3ad8093e49a7a5476d3