Fix more_like_api.feature
switch to -30% instead of 70%, this is basically the same thing but it works.
Spent the afternoon trying to figure out the machine running cindy fails on
this one. There must be something really strange happening there since the exact
same query works well locally.
What is strange is that running the query with mininum_should_match set to 99%
returns pages where only one term that match and setting 100% require all terms.

Change-Id: Iae64d730b48f7c83d6827e68a99b6f075153bec3