Profile management
Profile are accessible thanks to a SearchProfileService.
This service is loaded by collecting various SearchProfileRepositories.
A search profile repository is:
 - typed: what kind of profiles does it stores (similarity, rescores, functions...)
 - named

For now there are two main levels of profile repositories:
- cirrus_base: contains defaults profiles defined in cirrus code (does not
  depend on the config)
- cirrus_config: contains customzation made to existing wgCirrusSearch vars

The profile manager also includes the logic to determine the profile
to be used in a given context.

There are no detection of duplicates for few reason:
- I'd like to allow easy transitions to/from cirrus_base/cirrus_config
- This would require listing all the profiles stored in a repository.
  I prefer to avoid needing to extract a list to permit implementing
  other kind of repositories that may not be able to extract their content
  easily.

Extension can register to a hook to declare their own repositories.
This allows extensions to declare the same level of repositories:
- base (stored in the codebase)
- config (customizable by extension config vars)

We do not merge the repositories, we rely on php array which maintains insert
order. In case of name ambiguities the first repo that was declared wins.

Bug: T183279
Change-Id: Icca9a44373d26309ac4e1c2392cebecdf2fae46b