Add hook to override FulltextQueryBuilder
This is needed because the query builder class is defined by FT_QUERY
profile with current context in SearchContext, but there's no opportunity
for an extension to override that context.

Another way would be to make a hook to override the context, but this way
we get the default query builder for free and can reuse it in an appropriate
context.

Change-Id: Ibecea9f4f697e1788029adfbaa5e1e4ba28a6880