Add QueryStringRegexParser
Introduce a parser class that holds all the regexes needed to parse
the fulltext query into a simplified AST.
This parser is not yet plugged into production code and is basically
a noop since it needs the elastic query building code.

This new parser understands the boolean logic so it should allow
us to fix the double parsing problem:
- will allow to apply the boolean logic to special syntax
- won't fail on lucene QueryString syntax errors

The AST contains the needed nodes to parse what we currently
support:
- Boolean logic
- Words
- Phrases (including phrase prefix)
- FuzzyWord
- Wildcard word
- Prefix query (specialized wildcard)

Instead of trying to mimic perfectly what the old parser was doing
I decided to include a fixture file to describe what the parser should
do.
This breaks existing behaviors in some corner case but (I hope) will
allows us to have a clear definition of what the parser does. Some of
the fixture test are marked as "debatable" since it was not obvious
what solution to apply: e.g. how should we fix broken queries?

Wildcard handling has been completely reworked since it was (imo)
completely broken:
- w?t* currently generates: t? (title.plain:w*^20 OR all.plain:w*)

It now parses valid wildcards with the following limits:
- 3 wildcards allowed: fallbacks to simple words in case of a*b*c*d*
- must be non contiguous: hist*?y is not understood as a wildcard query.
- but word** is understood as a prefix

Bug: T188530
Change-Id: I061832c3516191c5110efef1d70cc36dd90f28c7