CirrusSearchCheckerJob should have a title
All the jobs provide a page title, some of them may even use
non-existent titles like "Special:Badtitle/JobSpecification" but the
checker job provides the empty string.

The new queue requires that the title is provided and since the
checker job is the only one that doesn't it should start doing that.

Bug: T190958
Change-Id: I8f305aca34a20e3c339f6bc96815c3def53e48ad