Introduce CrossSearchStrategy
Adds a new class to indicate the strategy to take when searching
external wikis.
Currently we support 3 types of crosswiki searches:
 - crossproject
 - crosslanguage (second try with language detection)
 - external indices (media files on commons)

From the usecases I've seen it's mostly allow everything or nothing
So a simple boolean might have been sufficient but I was worried
to introduce a simple boolean in method signatures. It'd be hard to
extent in the future.

Added a getCrossSearchStrategy( $node ) method the keyword features
so that a keyword can decide if it's able to run on external indices.

This might help us to support :
- the cost based decision to run crossproject searches
- the local keyword

Bug: T190132
Change-Id: I22a193ee36ee007f68f40681a735f139a832fbdd