Do not propagate Elastica doc modifications out of DataSender
Documents may be updated when running DataSender:sendData :
 - reset the noop hints
 - some metadata (_id, _version) by elastica when receiving the
   elastic response

But when writing on multiple clusters we want to send the same docs
to both clusters. Previously only the first cluster received the
original documents, subsequent clusters received a modified document
leveraging a sneaky bug in how we send the 'handlers' param to the
noop script.

Also fixed this edge case where the noop 'handlers' is empty to force
an empty object instead of an empty array that was causing
ClassCastException in the extra plugin.

Ideally we should also provide more information during error reporting
because the elastic exception was not reported in server logs nor MW
errors. We should probably inspect the response in the exception to
augment the data we log from MW.

Bug: T192609
Change-Id: I238fc3f64d54d149ac9c22a07f14f83ec187a785