Add KeywordFeature::expand
Adds a new method meant to be called just before query building.
It's meant to provide a method for keywords to apply costly process
such as calling external resources.
It'll be called on every target wikis the search request will be made
to.
The value generated by this method will be held in a context and
accessible when the keyword wants to build its query components.

Change-Id: I8a6b0642e607f20fbfe59b6a068a771485a8735b