Drop AfterImportPage hook subscriber
This appears to be duplicate work, the regular page edit
pipeline is triggered as well which inserts a LinksUpdate
job doing all the same things this MassIndex job does.

In addition there was a bug here, importing a redirect
would index the redirect directly instead of tracing
the redirect to the target.

Change-Id: I9df222ce58016e3cd322a852cc48e7c2769a80cd