Only issue delete on page move when necessary
This was issuing a delete whenever a page changed namespaces, even
if that was unnecessary. Add a check if the index suffix actually
changed and only delete when it has.

Deleting a page we want is generally a bad practice anyways, it makes
it temporarily unavailable for search and loses any external
properties such as a popularity_score. If the delete and index jobs
are re-ordered the page will potentially be missing until edited
or caught by the saneitizer.

Change-Id: I2c03cc3a5e0bdaade78f53fbbdcd3dfc41a090c4