Increase backoff time for retried jobs
In production we had the job drop timeout on freezes set to
three hours, but the job queue size didn't cap out it kept
climbing. This implies that the retries combined with three
hours of backlog was more than the job runners could process.

Help the job runners out by increasing the minimum delay to
two minutes on all but the first failure. Not sure this will
make a huge difference but it might help a bit.

Change-Id: I3c199dc0cb5e224ef284014d4e730c150083bf73