Prefix should not override the SearchContext namespace
It'll ease future developpement as this was one the sole feature
willing to update this kind of states in the SearchContext.
Instead of setting the namesapce at the SearchContext the keyword will
now provide the namespace filter inside its own filter.
A warning will be shown if the namespace found in the prefix keyword
is not compatible with the requested namespaces

Bug: T193392
Change-Id: I8ab8e9aa17bc7e036cdcf7b9135561d9036a73ec