Pass string flags in query parameters
When passing boolean query parameters into elastica they
get converted into 1/0. Elasticsearch used to be happy with
that but now they want everything to pass the strings 'true'
or 'false'.

It's plausible an update to Elastica would also fix this, but seems
easy enough to fix now and clear up some warnings.

Bug: T192614
Change-Id: I32c23209a02eaef4bd54bd30ba0cf30ed60c4b46