Pass around Printer interfaces instead of Maintenance
Not particularly important, but while looking at some other
code it seemed a little silly to pass the whole Maintenance
scripts in. Of course it's currently still a maint script,
but this makes it more obvious what it's for.

Additionally an early version of this patch broke all the
maintenance scripts without failing any unit tests
(integration tests failed). Add a test case that ensures
maintenance classes are all independantly loadable.

Change-Id: If509ced603a8101cb528ff1643741947073f72a1