Utilize EmptyResultSet where appropriate
We had this EmptyResultSet, but were returning a SearchResultSet
object from core to represent empty results in some places. Needed
to add the $containedSyntax parameter to the EmptyResultSet so
Special:Search can correctly offer to create pages.

This exposed that we were overriding searchContainedSyntax() from core
for no reason. Adjust our result set to call parent constructor with
appropriate args and stop reimplementing the getter.

Change-Id: I19a8d6621a130811871dec9335038797627d9448