Adjust available sort orders
The sort orders provided, specifically title_{asc,desc}, are very
expensive to support in elasticsearch and as such have never been
exposed to users. Adjust the list into something we can expose
to users, and update getValidSorts() to include those.

Additionally adds a new create_timestamp field to allow users to
sort by the creation date of pages. This is not exposed in
the SearchEngine interface yet due to the field needing to
be populated.

Bug: T195071
Change-Id: I745b2c463681cb68c16b9b18a07f79f3ba392d92