Force json float precision for test fixtures
It seems the default serialization precision of floats has changed
between hhvm 3.18 and php 7.2. Keep updates to the fixtures from
generating dirty diffs by forcing the precision to a known value.

Also does a small refactoring of fixture handling to centralize
access. It looks like there is quite a bit of duplicated code
and more refactoring could be done here.

Change-Id: I2cfb6cdf634c75ab83848fa563c9ea755b1b84aa