Prefix keyword should append its namespace not override the initial ones
Introduce a magic key in KeywordFeature::parseValue() results that the
parser can inspect to determine if a specific keywords requires
additional namespaces to function properly.

Added to support the behavior of the prefix keyword which allows to
set a namespace in its value.

Updated existing parsing behavior to append its namespace rather than
overridding it in the SearchContext.

Its query was already updated in prior patch to embark the namespace
filter as part of its filter query.

Bug: T195815
Change-Id: I975ddd8add98521c84b67a3cdda8b479fbb60b78