Add a fulltext query classifier
Introduce a mechanism to declare query classifiers.
A classifier works on top of the query AST.
The first classifier includes few types:
- simple_bag_of_words
- simple_phrase_query
- bag_of_words_with_phrase_query
- complex_query
- bogus_query

These query classes will be made available early in the process
so that search components can inspect them to decide whether or
not they can support a particular query.

This is meant to replace SearchContext::syntaxUsed in the future
by clearly separating:
- the query type (fulltext/prefix)
- the features used
- the classes of the query

Extensions can declare their own classifiers by registering to the
CirrusSearchRegisterFullTextQueryClassifiers hook.

Bug: T197774
Change-Id: I173b2de698f7fb9d515f1d50994f35852e07efad