Fix integration tests for V => Venom > V:N
This test was failing because it may happen that Venom
has 0 incoming_links (linked from Sam Wilson) this leads
to a very low score ~3000 but V:N obtains ~4000.
Increase text length for Venom so that even in the case
it lacks its incoming link it'll get around 6000.
Move Venom first and Sam Wilson last to have more chance
to set incoming_links before the suggester index is updated.

Bug: T199445
Change-Id: Id9aa8cfa1b32a626040daaec06f739460d34bb43