Cache cirrus queries to the WAN
Caching to only the local DC means that in a switchover the cache
hit rate goes straight to 0, and the cache is much too large and
expensive to generate to get by with a 10 or 20 minute warmup. Use the
WAN cache to separate cache hit rate from the dc being operated in.

Bug: T204148
Change-Id: Icaa653c47830d7e4cf649c848c268bf926184148