Start indexing the display_title property
The display title is used by things like the Translate extension
to replace the titles displayed on page. It would be useful
if CirrusSearch could find these. This patch is the first step which
will start indexing the data. Other patches which should be deployed
after this field is populated will add mappings and search
functionality.

Because display title is underdefined some heuristics for potentially
removing the namespace from the display title are used. This is
expected to have some edge cases but will hopefully improve the
experience when combined with the Translate extension.

Change-Id: I5208512edd0266563a9e3e316c3a9cce49335282