Include X-Search-ID header on api requests
When API requests perform a search inject an X-Search-ID
header into the response. This allows associating responses
with exact backend logs of the searches. The first use case will
be to record this id with autocomplete so processing can find
the result sets that were returned. This will then be used to
determine examination probabilities in autocomplete.

Bug: T205348
Change-Id: Ibc4bdb16db4c4be2817d4c888bd74d05adf12c92