Introduce SearchQuery and its builder
SearchQuery is a immutable object that holds all the necessary
information needed to:
- issue a query to the search backend
- determine which profile context to use
- check if a particular search fallback method should be applied

The profile context (which determines the set of profiles to use)
will be determined based on this data. A new component will be
added in up comming patches to let cirrus or extensions route
a particular query to a specific profile context.

Some bits are still a bit fuzzy at this point (ResultsType) as it's
still unclear where they belong to nor if it's necessary to make some
decisions about them so early.
The notion of "entry point" is a bit fuzzy as well. The aim is to
support all the search types but also in the future add a way
to tag carry some information from the UI/API to the query.
e.g. tell the query router that this query is issued from this
particular input box in the UI to trigger a specific set of profiles.
(the underlying question is: should the UI know the profile contexts
available and send specific params or should the searchEngine adapt
itself based on the UI using a kind of uiEntryPoint tag passed along)

NOTE: this starts using the new query parser, mostly to put the model
in place. It uses the following feature from the new parser:
- to determine the namespaces
- to determine the CrossSearch strategies

The goal will then be to gradually replace the old code using the
information available in SearchQuery/ParsedQuery.

Bug: T189880

Change-Id: Iba4584726974a3acecd61a12e4927e664409de50