Remove SearchConfig::getAvailableClusters
Remove this function and related functionality. It looks like
the only remaining use was to check if a cluster was valid
for use in saneitizeJobs, but that also has a check against
the set of writable clusters. Seems no need to keep this for
that minor distinction.

This is intended to help simplify what is exposed about
connections and clusters prior to adding in support for
elasticsearch cross-cluster search.

Change-Id: I53ce0e92ab37155bb841989ef88c34c659862d56