Really use target wiki config during cross project searches
this seemed wrong to use the host wiki config in Searcher::$config
while trying to build a query for another profject wiki.
Currently it seems that if we declare a rescore profile on the
target wiki that is not on the host wiki it will fail.
The reason is that the profile service will determine the default
rescore profile using SearchContext::getRescoreProfile() but will
then use the host wiki profile service to load it (Searcher::$config)

To fix this we need to reset Searcher::$config like what we do for
Searcher::$searchContext.

Change-Id: I808ec65c5d5aeb1a8f1184e29dcd452a91ee6ae3