Add NamespaceHeaderNode
Due to the way we "optimize" the AST on the fly we don't really know
when the actual query starts.
When querying -foobar we lose the offset of '-'.
We never needed to know this information up until now where we have
a mix of old and new code that needs to access the original query.
Rather than trying to change how we generate the AST (which is perhaps
the proper way to do) we introduce NamespaceHeaderNode which contain
the information we lack to determine the start of the query:
 - the query always starts at 0
 - or at the end of the namespace header: help:-foobar

Change-Id: Iae19ae1114812f5892c80e1c76ed559b8c0866a5