Updater: Pull only the timestamp of oldest revision
This query has triggered timeouts a few times in wmf prod.  By pulling
only rev_timestamp instead of the entire old revision this will
optimize out the lookup by primary key, all the info needed to answer
our query is in the rev_timestamp index already. I'm not 100% this
will solve the current issues seen in prod, but it will reduce the
amount of work done per query and we are constantly calling this code.

Bug: T206042
Change-Id: I1e2deedde4ed7933eea63122041f2cee0cac6625