Generalize fallback & second-try methods
Add a FallbackMethod to implement a fallback/retry strategy. 2 methods
added so far:
- phrase suggest for rewritting using the DYM suggestion
- language detection for appending new results using cross language
  search

The class FallbackRunner is responsible for running all these methods.
For now the default strategies are hardcoded:
- rewrite using DYM
- append if the previous results returned fewer results than the
  threshold set.

In the future it would be nice to extend the FallbackMethod to allow
it to amend the Query sent to elastic (move the code to build the
phrase suggest query into PhraseSuggestFallbackMethodTest), but also
move the code that extract the DYM from the elastic response.

Bug: T204255
Change-Id: I9f661d0b8d3618bff328f4fc31d3a874458c9a12