Properly test the "classic" profile name
Introduced in Ic9ff46b the problem was that we tested the profile array
instead of the profile name.
The "classic" was later fetched directly in the completion suggester
but because it's fake profile without any suggest options it generated
an invalid suggest request triggering an invalid json as per elastic
expectations.

Bug: T208248
Change-Id: Id2e12e61ad2a4c65fbe2393173096277e0dda513