Fix OtherIndex to comply with replica+group concepts
This was implemented in a way that does fit the new replica+group
concepts.
Simplify ExternalIndex handling by supporting a syntax similar to
cross cluster search. So that instead of defining 1 to 1 cluster
mappings we could simply define NS_FILE as "chi:commons_file".
Added a way to prevent writing OtherIndex data to some replica so that
during the transition period we can write wiki docs to psi but not the
accompanying OtherIndex updates.

Bug: T212120
Change-Id: Ibcc734c56df95c1625092d79884e14817692d211