Add QueryFixer
Is a simple visitor that tries to detect in a query what can be fixed
and the additional fix method to repair a query using an external
corrector (e.g. typo).
This class is meant to unify some of the logic around the phrase suggester.
It was designed to mimic as much as possible the current behavior excluding
incorrect incoherent ones.
We may not want to stick with these behaviors as I'm pretty sure most of them
were side-effect of the current parsing technique.
Dropped the support of merging two bags of words that are separated with a
keyword, I'm not sure it's right to consider them a "phrase". With a smarter
QueryFixer we could re-add this but not sure it's worthwhile.

Bug: T213098
Change-Id: Iabb63e794044378e80f12758c5692cf1fed324fe