Drop support for SourceRegex timeout param
Now that elastic has another way to monitor timeouts it's extremely
hard to tune it properly.
Reason is mostly because raising a timeout from the SourceRegex itself
will produce a search failure without giving a chance to see partial results.
The only way to see partial results is to hit the timeout within the elastic
check.
We ended having 3 timeouts to tune:
1/ request timeout: search failure when hit (regex potentially still running)
2/ elastic search request timeout: partial results when hit
3/ source regex timeout: search failure when hit

Since 1 & 3 would lead to no partial results I think we should just give
up on trying to be more precise in timeout checks.

Progress has been made on elastic side as it used to be only checked on segment
boundaries. It's now checked inside BulkScorer on a increasing 1024 docs interval.
Interval grows by a power of 2 meaning with a max at 2^20 which may be very long
if the regex is costly.

I think we should try to rely only on elastic timeout handling, if we see a rise
in request timeouts and long running queries server side we should revisit this
possibly sending PR to elastic to allow us to tune the bulk ID intervals so that
timeout checks are more frequent.

Bug: T198734
Change-Id: I6c580f9d7bcefc1f02bfb1fef726cd3638070fa0