Make all class names in @covers tags start with a backslash
Note that these tests are all organized in namespaces, e.g.
`namespace CirrusSearch`.

A line like `@covers CirrusSearch\DataSender` that does not start with a
backslash is technically a *relative* class name, relative to the current
namespace. Tools and IDEs like my PHPStorm think they have to look for a
class \CirrusSearch\CirrusSearch\DataSender, which doesn't exist.

It's good practice to use absolute, full qualified class names in @covers
tags. This codebase demonstrates why this is. :-)

Note: The results at https://doc.wikimedia.org/cover-extensions/CirrusSearch/
show that the PHPUnit version we are using to generate our coverage reports
does not care, and is able to understand these @covers tags anyway. But not
all tools are like this.

Change-Id: I10cdc293c07c1c35d74ca0af94c80fe2115337d0