Add a search query dispatch service
Meant to let extension control the profile context based on a set of
custom heuristics (namespaces/query features).
This is meant to replace the hooks CirrusSearchFulltextQueryBuilder
and CirrusSearchFulltextQueryBuilderComplete.
For now cirrus provides a "static" routing mechanism but extensions
are free to implement their own routes with varying scores depending
on the information available in SearchQuery.
Cirrus will always provide a "default route" using its classic query
builders which are guarantee to "work" (search and provide results) on
all indices managed by cirrus.

Bug: T216429
Change-Id: I9e5d3751b4c989c064f23145668ba34368422198