Add FetchPhaseConfigBuilder
Add builder object to facilitate the construction of the various
fetch phase elements in the query.
Only supports highlighting for now but will be extended to support
other elements (stored fields and source filtering).
Only fulltext has been migrated but prefix will follow.
The overall idea is to start tracking what is the "target" of an
element we fetch from elastic (hence the notion of priority because
in some cases there are competing source data for a single target).
A target as defined now means to be a place (field) in the Result instance.
This particular patch does not include the generic transformation
mechanism. In other words the state of the highlight configuration
indirectly reflects the priority mentioned earlier (the reason few hacks
are still needed).

The sole change in the highlight config should be that we now continue
to highlight all other fields even if a regex is used.

Bug: T190130
Change-Id: I9e0b768d55f59eb82d1536e1b39557f33beb07f2