Remove integration test for Special:Version
This is testing that Special:Version is working correctly rather
than CirusSearch. The presence of this row is controlled by
extension.json which is declarative. I'm not sure every extension
needs to hardcode the framework of this core special page to
test that that works correctly.

The one thing it does often is a proxy for is that there are no
fatal errors during installation. This is useful when less-savvy
administrators install MediaWiki and might not recognise errors
or non-zero exit codes on the command-line, and thus checking
Special:Version is a common way for end-users to verify that it
worked.

However, in CI we already fail on all such scenarios which is
very much its own integration test effectively already for anything
happening during the installation process.

Removal motivated by similar tests being flaky in existing
extensions (T220758), or recommended against (by me) for new
tests (I8c53d9c923820).

Change-Id: Ib6917f827ce90cf549a87aa2df81094f9ed50974