Pass connection to ElasticsearchIntermediary::succcess/failure
The recent usage of cluster overrides to deal with weekend overloads
has broken one of the failure cases. Pass the appropriate connection
object through to ensure we always have access to the last performed
request.

While at it fix another bug in the same code exposed by the same
overrides, which assigns latency statistics to the original cluster
instead of the cluster overridden to.

Bug: T221058
Bug: T221019
Change-Id: I8560baa69965d46318a9bc3f580363b3655aa800