Don't validate request event schema
It turns out mw-core has justinrainbow/json-schema, and
cirrus is (err, was) using swaggest/json-schema. Additionally
the YAML dependency is not appropriately available for extensions
that depend on cirrus.

We should bring this test back, but it needs to be better aligned
with the other dependencies and we need to decide how the yaml
library will also get to all those other tests. Likely this needs
to be reworked to use justinrainbow, and the yaml dep needs to
be added to mediawiki-core require-dev.

That all will take a little longer though, and right now we need
to unbreak the build.

Bug: T221484
Change-Id: Idaf0c812069703ecbad0843ca652fe297127e43f