Allow per-cluster CirrusSearchMaxShardsPerCluster
If clusters being served are very different in sizing they may need
different values for the max shards per cluster. In particular a
production search cluster may want to use 1 to ensure shards spread
across a large number of nodes, while a small replica for non-prod
usage may want 'unlimited' as there are not enough nodes to spread
the shards around. Ideally the shard counts should be reduced to
better match the cluster, but there are limits to how large a single
shard can be.

Bug: 220625
Change-Id: I459e3319b0fd5d99ba206e314a314ede4a1a5e4d