Limit the clusters archive index is written to
The archive index contains private data that cannot be exposed
directly to users. Provide the ability to create a whitelist of
clusters that are allowed to contain private indices. This will
allow maintaining public replicas of the search indices without
exposing private data to them.

A more generic implementation, which allowed configuring indices
available on various clusters, was explored but deemed more
complicated than we want for this single use case. Instead this
is hard coded to the archive index.  Adding new indices is quite rare,
but if in the future we add additional private indices this should be
revisited.

Bug: T220625
Change-Id: I246393ae3b704d354ff9364df18ee82dd9d0145d