Protect HashSearchConfig from incorrect flags
Make constants for all flags accepted by HashSearchConfig, and
validate that all flags provided are in the set of known flags.

Job::decideCluster had 'inherited' instead of 'inherit' for the flag
which caused a non-obvious error when the config object returned null
for requested values. Push the error further up the stack by
validating early.

Bug: T222819
Change-Id: I7fda1b9acde3a337112e395ad576e10303b039cd