Clarify Connection usage in Job/Updater classes
remove the connection member from the abstract Job class it was unclear
whether it could be use for writing (to which cluster?) or reading.
Also remove the $flags in Updater it was only used for the 'same-cluster'
semantic. Prefer passing the elastic connection that will be used for
reading and the cluster name to be written to (null for all).

Change-Id: I16277ba2b44bba5fcd42259fe9d641cbcfca197b