Validate request logging with json schema
This is a partial revert of 251c1958 which was the first attempt
at validating the event schema in the cirrus test suite.

This exposes that the search_token field needs to be
renamed to search_id to match the schema. Verified in hadoop that
this value is missing in the stored events.

Additionally the schema did not match the mediawiki/event-schemas
repository. The schema fixture has been updated to match.

Bug: T222268
Depends-On: Iddbfb81fb667980470d6872c34a1e63897f1e4af
Change-Id: I5488e0e61277c42e21e81f91fc6c9705c7fdfdd3