Don't spam the logs with errors from saneitizer jobs when elastic is down
During cluster restarts some queries to fetch the ids to check may
fail. This job was relying on the job queue internal retry technique
but at the cost of log pollution (exception thrown). Since the
saneitization process is not critical it's not worthwhile to alert
about these failures with the ERROR severity, simply reuse our
internal retry mechanism and prevent the exception to propagate to
the core job runner.

Bug: T222921
Change-Id: I188cca1ec098eb22586c54cdd0998fbf7e42e855