RedirectsAndIncomingLinks: succeede or fail, but not both
Due to the placement of the try/catch and success call we could mark
the request as a success, and then throw an exception and try to
turn it into a failure. This results in trying to close a request that
was never opened, and an NPE.

Additionally fix the call site in ElasticsearchIntermediary::failure
that broke due to this being called twice. We are already in a failure
condition, the fact that the request wasn't open has been logged. We
shouldn't really do anything more than continue on in this case.

Bug: T227691
Change-Id: I3f962875f70e5b715ffe90cc90dae20be8588487