build: temporarily rename 'selenium-test'
Wikimedia CI uses NodeJS 10. wdio-cucumber-framework depends on
fibers@2.x which is not compatible.  We have to bump it to support
NodeJS 10 but then another CI system (Cindy the browser bot) is still on
NodeJS 4 or 6 and would thus not work anymore.

Since Quibble 0.0.34 look up for a 'selenium-test' script, rename it
temporarily to hide the script from Quibble. This way it will no more
attempt to npm install fibers@2.x and break.

The selenium tests are still run from mediawiki/core which has a
wildcard pattern for extensions/*/tests/selenium/specs/**/*.js.

Bug: T229033
Change-Id: Ic1be67c4231e0b0d6865f6fa1209c3354d2ff43f