Add CirrusSearchResultBuilder and ArrayCirrusSearchResult
ArrayCirrusSearchResult is meant to be a flexible replacement of the
Result class. CirrusSearchResultBuilder is an helper class allowing to
create ArrayCirrusSearchResult instances in a type safe way.
Also adds FullTextCirrusSearchResultBuilder which wraps CirrusSearchResultBuilder
to read and transform a elastica Result into an ArrayCirrusSearchResult
that is equivalent to one created using the Cirrus Result constructor.

End goal of these classes is to take an Elastica Result instance and
transforms it into an ArrayCirrusSearchResult using CirrusSearchResultBuilder
while being driven by the configuration set in FetchPhaseConfigBuilder.

Had to make TitleHelper a injectable component to ease testing.

Bug: T190130
Change-Id: I6651fb49d9e54738e2db314663dc631fdbfff262