Add morelikethis a non-greedy version of the morelike keyword
morelike is a keyword that cannot be combined because it is greedy
and considered a query header. This does not allow to combine it
with any other keywords/query text.
Since morelike is heavily used it is not possible to fix without
breaking existing clients.
Introduce a new keyword "morelikethis" to address all the limitations
of morelike.
Like any other keywords its value must be wrapped in double quotes if
it contains space.
Internally it will act as a more_like feature:
- be reported as such in metrics
- cached using the same TTL config

Bug: T159321
Change-Id: I81bda72fa7fb4f544d4bed701d441a473d00329e
(cherry picked from commit cd8ef9520c06c0ec396bd18c1cf4efa189869972)