Ignore bad (UTF8) links returned from LinksUpdate::getRemovedLinks
Most likely stored in the pagelinks table when the parser was still
allowing such links to be parsed (c.f. T13143).
When such old pages are updated (either directly or indirectly)
LinksUpdate will notify us that these links are removed. CirrusSearch
should simply ignore such links allowing the job to be pushed to the
jobqueue without errors when encoding it to json.

Bug: T230728
Change-Id: I9065b2522c9c229ce752e5484ee43b04e86b6d23