Start using priority of highlighted fields
As of now the response was only needed to figure out what field
from the elasticsearch response needs to be taken for a particular
target in the SearchResult.
FullTextResultsType will now pass the configuration built using
FetchPhaseConfigBuilder which contains a notion of priority.
It is far for being fully flexible but should allow keyword to interact
with the highlighting behaviour if they hook to a supported target.
ResultSet is being deprecated and Result will follow.
Moved dependent integration tests to the new unit tests. This caused
some additional changes (move more dependencies to constructors).

Bug: T190130
Change-Id: I03e1409bdca4db2b08d0cf3f3561dc5ade8f407b
Depends-On: I1b67a9cddfca7fa06fb0b916981d1efde0764018