Make sure to use host wiki components when issuing a sister wiki search
The original bug was in Searcher::makeSearcher it should have reused
the TitleHelper instance passed to the original Searcher.
Also do not depend on SearchConfig just for accessing the wiki id.
Detecting if the config is targeting a sister or the host wiki is not
yet entirely clear yet and some refactoring are still needed to completely
remove global states access from SearchConfig/HashSearchConfig.

Bug: T235778
Change-Id: Ib86ece0fff9549ad3edf255bce53ac7ccf1bb3dd